Fixed x64 jmp patching

+++ Jit_Compiler.h
@@ -128,8 +128,8 @@ class JitCompiler_t
         if (sizeof (T) == sizeof (int64_t))
         {
             //ErrorPrintfBox ("%d %s", __LINE__, __PRETTY_FUNCTION__);
-            man.EmitPush<int32_t> ((int32_t)((int64_t)imm >> (sizeof (int32_t)*8)));
-            man.EmitPush<int32_t> ((int32_t)((int64_t)imm));
+            man.EmitPush<int32_t> (static_cast<int32_t>(static_cast<int64_t>(imm) >> (sizeof (int32_t)*8)));
+            man.EmitPush<int32_t> (static_cast<int32_t>(static_cast<int64_t>(imm)));
         }
         else man.EmitPush<T> (imm);
     }
